# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "This is an open API for `BAD` operations management. You can use the\
    \ api key `api_key` to test the authorization filters. <h2>Authentication</h2>\
    \ All API's must be authenticated. To authenticate, supply your API key as the\
    \ `Authorization` HTTP header with the `api_key` authorization scheme. <h2>Parameters</h2>\
    \ For API's that use POST, set the request's Content-Type header to `application/json`\
    \ and supply an appropriate JSON object in the request body. For API's that use\
    \ GET, specify each request parameter as a properly-escaped query parameter. <h2>Response</h2>\
    \ All responses have their Content-Type set to `application/json`, and contain\
    \ the response JSON object in the response body."
  version: "1.0.0"
  title: "BAD and Certified pick-ups OpenAPI"
host: "localhost:8081"
basePath: "/v2"
...
schemes:
- "http"
consumes:
- "application/json"
paths:
...
  /createBad:
    post:
      tags:
      - "Bad"
      summary: "Create a new bad in the blockchain."
      description: "Issues a new bad to the world state with the given details.\n"
      operationId: "createBadPOST"
      parameters:
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/BadWithSignature"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /readBad:
    get:
      tags:
      - "Bad"
      summary: "Read a bad stored in the blockchain with the given badId."
      description: "Retrieves the bad stored in the blockchain with the given badId.\n"
      operationId: "readBadGET"
      parameters:
      - name: "badId"
        in: "query"
        description: "ID of the bad"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/bad"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /deleteBad:
    delete:
      tags:
      - "Bad"
      summary: "Delete a given bad from the world state."
      description: "Deletes the bad with the given badId from the world state on the\
        \ blockchain.\n"
      operationId: "deleteBadDELETE"
      parameters:
      - name: "badId"
        in: "query"
        description: "ID of the bad"
        required: true
        type: "string"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /badExists:
    get:
      tags:
      - "Bad"
      summary: "Check if bad with given ID exists in world state."
      description: "Checks if the bad with the given badId exists in the blockchain.\n"
      operationId: "badExistsGET"
      parameters:
      - name: "badId"
        in: "query"
        description: "ID of the bad"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /changeOwner:
    put:
      tags:
      - "Bad"
      summary: "Update the owner AND CONSIGNEE field of a bad with given badId in\
        \ the blockchain."
      description: "Updates both the owner and consignee fields of a bad with the\
        \ given badId in the blockchain.\n"
      operationId: "changeOwnerPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/BadChangeOwnerRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /changeStatus:
    put:
      tags:
      - "Bad"
      summary: "Update the Status of a bad with given badId in the blockchain."
      description: "Update the Status of a bad with given badId in the blockchain.\n"
      operationId: "changeStatusPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/BadChangeStatusRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /changeStatusWithData:
    put:
      tags:
      - "Bad"
      summary: "Update the bad status and other details given a badId in the blockchain."
      description: "Update the bad status and other details given a badId in the blockchain.\n"
      operationId: "changeStatusWithDataPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/BadChangeOwnerAndStatusRequestWithoutSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /changeHolder:
    put:
      tags:
      - "Bad"
      summary: "Update the holder field of a bad with given badId in the blockchain."
      description: "Updates the holder field of a bad with the given badId in the\
        \ blockchain.\n"
      operationId: "changeHolderPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/BadChangeHolderRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /endorseTransfer:
    put:
      tags:
      - "Bad"
      summary: "Update the holder AND owner field of a bad in the blockchain."
      description: "Updates both the holder and owner fields of a bad in the blockchain.\n"
      operationId: "endorseTransferPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/EndorseTransferRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /resbmitBad:
    put:
      tags:
      - "Bad"
      summary: "Update bad after customs response."
      description: "Updates bad fields in the blockchain.\n"
      operationId: "resubmitBadPUT"
      parameters:
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/ResubmitRequestWithSignature"
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /getAllBads:
    get:
      tags:
      - "Bad"
      summary: "Return all bads found in the blockchain."
      description: "Retrieves a list of all bads found in blockchain. The results\
        \ are pagginated with pagesize.\n"
      operationId: "getAllBadsGET"
      parameters:
      - in: "body"
        name: "body"
        description: "Pegination details"
        required: true
        schema:
          $ref: "#/definitions/getallRequest"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/inline_response_200"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /getFiltredBads:
    get:
      tags:
      - "Bad"
      summary: "Return all bads found in the blockchain based on defined filters."
      description: "Retrieves a list of all bads found in blockchain. \n"
      operationId: "getfiltredBadsGET"
      parameters:
      - in: "body"
        name: "body"
        description: "Pegination details"
        required: true
        schema:
          $ref: "#/definitions/getfiltredBadsRequest"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/inline_response_200"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /getBadByOwner/:
    get:
      tags:
      - "Bad"
      summary: "Return all bads found in the blockchain."
      description: "Retrieves a list of all bads owned by a given user.\n"
      operationId: "getBadsByOwnerGET"
      parameters:
      - name: "ownerName"
        in: "query"
        description: "Pegination details"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/dateInterval"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/bad"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /getBadByTerminal/:
    get:
      tags:
      - "Bad"
      summary: "Return all bads found in the blockchain by Terminal."
      description: "Retrieves a list of all bads owned by a given terminal.\n"
      operationId: "getBadByTerminalGET"
      parameters:
      - name: "terminal"
        in: "query"
        description: "Pegination details"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        required: false
        schema:
          $ref: "#/definitions/dateInterval"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/bad"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /breakBad:
    post:
      tags:
      - "Bad"
      summary: "Transform a bad into multiple CPUs based on NumCpu."
      description: "Transforms a bad into multiple CPUs based on the value of NumCpu\
        \ field in the bad object.\n"
      operationId: "breakBadPOST"
      parameters:
      - name: "initiator"
        in: "query"
        description: "Inintiator of the transaction"
        required: true
        type: "string"
      - in: "body"
        name: "body"
        description: "BAD object"
        required: true
        schema:
          $ref: "#/definitions/BreakBadRequestWithSignature"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
...
  /isHolder:
    get:
      tags:
      - "Bad"
      summary: "Check if the user is the holder of the specified bad in the blockchain."
      description: "Checks if the user is the holder of the specified bad in the blockchain.\n"
      operationId: "isHolderGET"
      parameters:
      - name: "badId"
        in: "query"
        description: "ID of the bad"
        required: true
        type: "string"
      - name: "holder"
        in: "query"
        description: "Holder of the bad"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
  /isOwner:
    get:
      tags:
      - "Bad"
      summary: "Check if the user is the owner of the specified bad in the blockchain."
      description: "Checks if the user is the owner of the specified bad in the blockchain.\n"
      operationId: "isOwnerGET"
      parameters:
      - name: "badId"
        in: "query"
        description: "ID of the bad"
        required: true
        type: "string"
      - name: "owner"
        in: "query"
        description: "Owner of the bad"
        required: true
        type: "string"
      security:
      - api_key: []
      responses:
        "200":
          description: "Success response"
          schema:
            $ref: "#/definitions/Message"
        "400":
          description: "Bad request"
          schema:
            $ref: "#/definitions/Message"
        "404":
          description: "Bad not found"
          schema:
            $ref: "#/definitions/Message"
        "500":
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Message"
      x-swagger-router-controller: "Bad"
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
definitions:
...
  getallRequest:
    type: "object"
    properties:
      pageSize:
        type: "integer"
      bookmark:
        type: "string"
  getfiltredBadsRequest:
    type: "object"
    properties:
      dateCreationDebut:
        type: "string"
      dateCreationFin:
        type: "string"
      numeroEscale:
        type: "string"
      numeroConnaissement:
        type: "integer"
      statusDouane:
        type: "string"
      numeroDeclarationSommaire:
        type: "string"
      numeroBAD:
        type: "string"
  dateInterval:
    type: "object"
    properties:
      startDate:
        type: "integer"
        description: "starting unix timestamp"
      endDate:
        type: "integer"
        example: 12324324
        description: "ending unix timestamp"
...
  BadChangeOwnerAndStatusRequestWithoutSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/BadChangeOwnerAndStatusRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        badId: "badId"
        newHolder: "newHolder"
        newOwner: "newOwner"
        status: "status"
...
  BadChangeOwnerRequestWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/BadChangeOwnerRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        badId: "badId"
        newOwner: "newOwner"
...
  BadChangeStatusRequestWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/UpdateStatusRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        badId: "badId"
        status: "status"
...
  EndorseTransferRequestWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/EndorseTransferRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        badId: "badId"
        consignee: "consignee"
        lieuStockage: "Term2"
        paiementConditional: true
        dateExpiration: 0
        issuer: "FRS"
        notify: "alami"
        status: "SUBMIT"
        montantBC: 24545
  ResubmitRequestWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/ResubmitRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        badId: "badId"
        consignee: "consignee"
        lieuStockage: "Term2"
        paiementConditional: true
        dateExpiration: 0
        issuer: "FRS"
        notify: "alami"
        status: "SUBMIT"
        montantBC: 24545
...
  BadChangeHolderRequestWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/BadChangeHolderRequest"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        badId: "badId"
        newHolder: "newHolder"
  bad:
    type: "object"
    required:
    - "badDetails"
    - "badId"
    - "consignee"
    - "issuer"
    - "owner"
    properties:
      badId:
        type: "string"
      issuer:
        type: "string"
      consignee:
        type: "string"
      owner:
        type: "string"
      montantBC:
        type: "integer"
        description: "Amount in centimes. Expecting an integer not a a float"
      holder:
        type: "string"
      notify:
        type: "string"
      creationDate:
        type: "integer"
        description: "TimeStamp"
      paiementConditional:
        type: "boolean"
      dateExpiration:
        type: "integer"
      isPaid:
        type: "boolean"
      numCpu:
        type: "integer"
        description: "how many cpus the bad should generate once broken"
      status:
        type: "string"
        enum:
        - "SUBMIT"
        - "VALID"
        - "REJECTED"
      badDetails:
        $ref: "#/definitions/BadDetails"
    example:
      owner: "owner"
      consignee: "consignee"
      holder: "holder"
      creationDate: 6
      issuer: "issuer"
      notify: "notify"
      montantBC: 0
      badDetails:
        dateCreation: "2023-10-13 15:39:06"
        notifyBad: "notifyBad"
        tar: 2000
        dateReceptionEai: "dateReceptionEai"
        poids: 2000
        voyage: "{}"
        referenceDeclarationSommaire: "ref_230000107"
        referenceBAD: "411230000000077"
        submitted: true
        idNotify: "195450"
        lieuStockage: "411001"
        connaissement: "{}"
        port: "MAPTM"
        freightForwarder: true
        dateReceptionDouane: "dateReceptionDouane"
      isPaid: true
      badId: "badId"
      paiementConditional: true
      dateExpiration: 1
      numCpu: 5
      status: "SUBMIT"
  signaturePayload:
    type: "string"
    description: "Signature of the body payload in base64"
    example: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
  BadWithSignature:
    type: "object"
    properties:
      bad:
        $ref: "#/definitions/bad"
      signaturePayload:
        $ref: "#/definitions/signaturePayload"
    example:
      signaturePayload: "C65xH0wn/uJOUyXnTOTdyafwL7JF4VbIjPyBLL77LjBqY7ltGeUTFcyvW6tolv8VGe+xhDDofRt2BAsiVylX8CNEih3GebDQ8thY8zEsiFVXCmYrll+AK7/MxfoTJ4QyKY/amGNKhPSdD4rYMZU4gKxFRKwsGAMoJeW+kP1ZSb0MLtbi11KTbyC1wYTfsnIa3+5PlfEbocMQQOLVwPSIcW2F66+1zyiHC4qQmM4Tr+jqblkXlTMWk1ptjqpByP6K3ASbCEQ3DHlGLFE1WTf8PpVdHJ3IWdus2hPoPKLpCXt9fNveCURy2iEPAcNA6bjwy8ptlcLpNM5nUgoTEsbrLg=="
      bad:
        owner: "owner"
        consignee: "consignee"
        holder: "holder"
        creationDate: 6
        issuer: "issuer"
        notify: "notify"
        montantBC: 0
        badDetails:
          dateCreation: "2023-10-13 15:39:06"
          notifyBad: "notifyBad"
          tar: 2000
          dateReceptionEai: "dateReceptionEai"
          poids: 2000
          voyage: "{}"
          referenceDeclarationSommaire: "ref_230000107"
          referenceBAD: "411230000000077"
          submitted: true
          idNotify: "195450"
          lieuStockage: "411001"
          connaissement: "{}"
          port: "MAPTM"
          freightForwarder: true
          dateReceptionDouane: "dateReceptionDouane"
        isPaid: true
        badId: "badId"
        paiementConditional: true
        dateExpiration: 1
        numCpu: 5
        status: "SUBMIT"
...
  Message:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        minimum: 0
      message:
        type: "string"
      error:
        type: "boolean"
    example:
      code: 0
      error: true
      message: "Response Message"
...
  inline_response_200:
    type: "object"
    properties:
      bookmark:
        type: "string"
        example: "|bad|badId186|"
      result:
        type: "array"
        items:
          $ref: "#/definitions/bad"
    example:
      result:
      - owner: "owner"
        consignee: "consignee"
        holder: "holder"
        creationDate: 6
        issuer: "issuer"
        notify: "notify"
        montantBC: 0
        badDetails:
          dateCreation: "2023-10-13 15:39:06"
          notifyBad: "notifyBad"
          tar: 2000
          dateReceptionEai: "dateReceptionEai"
          poids: 2000
          voyage: "{}"
          referenceDeclarationSommaire: "ref_230000107"
          referenceBAD: "411230000000077"
          submitted: true
          idNotify: "195450"
          lieuStockage: "411001"
          connaissement: "{}"
          port: "MAPTM"
          freightForwarder: true
          dateReceptionDouane: "dateReceptionDouane"
        isPaid: true
        badId: "badId"
        paiementConditional: true
        dateExpiration: 1
        numCpu: 5
        status: "SUBMIT"
      - owner: "owner"
        consignee: "consignee"
        holder: "holder"
        creationDate: 6
        issuer: "issuer"
        notify: "notify"
        montantBC: 0
        badDetails:
          dateCreation: "2023-10-13 15:39:06"
          notifyBad: "notifyBad"
          tar: 2000
          dateReceptionEai: "dateReceptionEai"
          poids: 2000
          voyage: "{}"
          referenceDeclarationSommaire: "ref_230000107"
          referenceBAD: "411230000000077"
          submitted: true
          idNotify: "195450"
          lieuStockage: "411001"
          connaissement: "{}"
          port: "MAPTM"
          freightForwarder: true
          dateReceptionDouane: "dateReceptionDouane"
        isPaid: true
        badId: "badId"
        paiementConditional: true
        dateExpiration: 1
        numCpu: 5
        status: "SUBMIT"
      bookmark: "|bad|badId186|"
...